requirements.txt
torch
transformers
accelerate
numpy
tqdm
scikit-learn
matplotlib
pandas
umap-learn
plotly
nltk
